requirements.txt
ffmpeg_python==0.2.0
numpy==2.2.6
soundfile==0.13.1
torch==2.7.0
transformers==4.51.3
whisper==1.1.10
